requirements.txt
appdirs==1.4.3
Django==1.11
django-appconf==1.0.2
django-compressor==2.1.1
django-libsass==0.7
ez-setup==0.9
libsass==0.13.4
olefile==0.44
packaging==16.8
pyparsing==2.2.0
pypugjs==5.6.*
pytz==2017.2
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
django-precise-bbcode>1.2.9
django-ckeditor>=5.6.1
django-embed-video>=1.2.0
Markdown>=3.0.*